@@ -25,14 +25,14 @@
   "devDependencies": {
     "babel-cli": "^6.22.2",
     "babel-core": "^6.22.1",
-    "babel-jest": "^19.0.0",
+    "babel-jest": "^20.0.3",
     "babel-loader": "^6.2.10",
     "babel-plugin-lodash": "^3.2.11",
     "babel-polyfill": "^6.22.0",
     "babel-preset-es2015": "^6.22.0",
-    "babel-preset-react": "^6.22.0",
-    "enzyme": "^2.3.0",
-    "enzyme-to-json": "^1.4.5",
+    "babel-preset-react": "^6.24.1",
+    "enzyme": "^2.8.2",
+    "enzyme-to-json": "^1.5.1",
     "eslint": "^2.9.0",
     "eslint-config-wagtail": "0.1.1",
     "eslint-import-resolver-webpack": "^0.8.1",
@@ -47,22 +47,24 @@
     "gulp-sourcemaps": "~1.5.2",
     "gulp-util": "~2.2.14",
     "imports-loader": "^0.6.5",
-    "jest": "^19.0.0",
+    "jest": "^20.0.4",
     "mustache": "^2.2.1",
     "react-addons-test-utils": "^15.4.2",
+    "react-test-renderer": "^15.5.4",
     "redux-mock-store": "^1.2.2",
     "require-dir": "^0.3.0",
     "webpack": "^1.12.14"
   },
   "dependencies": {
-    "focus-trap-react": "^3.0.2",
+    "focus-trap-react": "^3.0.3",
     "lodash": "^4.17.4",
-    "react": "^15.4.2",
-    "react-addons-css-transition-group": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.5",
+    "react-transition-group": "^1.1.3",
     "redux": "^3.6.0",
-    "redux-actions": "^1.2.1",
+    "redux-actions": "^2.0.3",
     "redux-thunk": "^2.2.0",
     "whatwg-fetch": "^2.0.2"
   },
