@@ -42,7 +42,8 @@ module.exports = function exports() {
       loaders: [
         {
           test: /\.js$/,
-          loader: 'babel'
+          loader: 'babel',
+          exclude: /node_modules/,
         },
       ]
     },
